+++ packages/docmap/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @paleo/docmap
 
+## 0.5.1
+
+### Patch Changes
+
+- Fixed escaping directory.
+
+## 0.5.0
+
+### Minor Changes
+
+- 18f400e: Positional path arguments replace `--dir` / `--read`.
+
 ## 0.4.4
 
 ### Patch Changes